Add video to image sequence, image sequence to gif
video_to_image_sequence() is a new function that
converts a video to an image sequence

convert_gif() now can convert image sequences to gifs

changed comments/organization of converter.py to be more clear

add new acceptable_extension() function with longer list of
acceptable file extensions

+++ converter.py
@@ -15,10 +15,13 @@ import re
 import shutil
 from natsort import natsorted
 
+import pathlib
+
 current_path = os.path.abspath(os.getcwd())
 input_path = current_path
 
-""" IMAGE OR GIF ALTERING FUNCTIONS """
+""" --- IMAGE OR GIF ALTERING FUNCTIONS --- """
+""" 1. FUNCTIONS THAT ALTER STATIC IMAGES """
 
 def posterize(filename, output_name, keep_transparency = True, bits = 2):
     """
@@ -194,24 +197,84 @@ def pillow_adjust_image(filename, output_name, new_saturation = 1.0, new_contras
     inp = ImageEnhance.Sharpness(inp).enhance(new_sharpness)
     inp.save(output_name)
 
+""" 2. FUNCTIONS THAT TURN VIDEOS INTO IMAGE SEQUENCES """
+
+def video_to_image_sequence(filename, output_name, fps = 10):
+    """ 
+    Turns input video into output png sequence.
+    Outputs are named output_name_1.png, output_name_2.png, etc.
+
+    fps = frames of the video to turn into pngs each second.
+    """
+
+    if not filename.endswith((".qt", ".mp4", ".mov")):
+        return
+
+    command = "ffmpeg -i " + filename + " -vf \"fps=" + str(fps) + "\" " + output_name + "_%d.png"
+    subprocess.call(command, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL, shell = True)
+
+""" 3. FUNCTIONS THAT TURN FILES INTO GIFS """
+
 def convert_gif(filename, output_name, number_frames = 8, framerate = 8): 
     """ 
     Turns input png into output gif.
     The gif will appear identical to the input png (it will not be animated).
+    
+    Also works on png sequences. If the input png ends with a number, e.g. string1.png, then any png starting with string will be incorporated into the resulting gif. To prevent creating the same gif twice, the gif will only be created if the input png sequence has a creation date newer than the resulting gif (if the resulting gif already exists).
 
-    Mainly useful for the painting function, which only converts pngs to pngs and gifs to gifs.
+    Useful for the painting function, which only converts pngs to pngs and gifs to gifs.
 
-    number_frames = number of frames in resulting gif.
+    number_frames = number of frames in resulting gif. Ignored with png sequence input.
     framerate = framerate of resulting gif.
     """
 
-    shutil.rmtree("temp", ignore_errors = True) 
-    os.makedirs("temp")
-    for i in range(number_frames):
-        shutil.copyfile(filename, os.path.join("temp", str(i) + ".png"))
+    # from https://stackoverflow.com/a/14471236
+    filename_no_extension = pathlib.Path(filename).stem
+    m = re.search(r'\d+$', filename_no_extension)
+    # if the filename ends in digits, m will be a Match object; otherwise it will be None
+    
+    # if input is png sequence
+    if m is not None:
+        common_prefix = filename_no_extension[:-len(m.group())]
+        frame_min = float("inf")
+        frame_max = float("-inf")
 
-    command = "ffmpeg -y -framerate " + str(framerate) + " -i temp/%d.png " + output_name + ".gif"
-    subprocess.call(command, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL, shell = True)
+        real_output_name = re.sub(r'\d+$', '', pathlib.Path(output_name).stem) + ".gif"
+        output_gif_exists = os.path.exists(real_output_name)
+        if output_gif_exists:
+            output_gif_creation = os.path.getmtime(real_output_name)
+            latest_sequence_creation = 0
+
+        # get frame_min and frame_max
+        for path in sorted(os.listdir(current_path)):
+            if path.startswith(common_prefix):
+                
+                m = re.search(r'\d+$', pathlib.Path(path).stem)
+                frame_num = int(m.group())
+                
+                frame_min = min(frame_min, frame_num)
+                frame_max = max(frame_max, frame_num)
+               
+                if output_gif_exists and os.path.getmtime(path) > latest_sequence_creation:
+                    latest_sequence_creation = os.path.getmtime(path)
+    
+        if not output_gif_exists or latest_sequence_creation > output_gif_creation:
+            command = 'convert $(printf -- "-dispose Background "; for ((a=' + str(frame_min) + '; a<' + str(frame_max) + '; a++)); do printf -- "-delay ' + str(framerate) + ' ' + common_prefix + '%d.png -dispose Background " $a; done;) ' + real_output_name
+           
+            # From https://stackoverflow.com/a/29269316 
+            # Normal subprocess.call or subprocess.run doesn't work here, possibly because of the nested commands
+            process = subprocess.Popen('/bin/bash', stdin=subprocess.PIPE, stdout=subprocess.PIPE, text=True)
+            out, err = process.communicate(command) 
+       
+    # if input is single png file 
+    else:
+        shutil.rmtree("temp", ignore_errors = True) 
+        os.makedirs("temp")
+        for i in range(number_frames):
+            shutil.copyfile(filename, os.path.join("temp", str(i) + ".png"))
+
+        command = "ffmpeg -y -framerate " + str(framerate) + " -i temp/%d.png " + output_name + ".gif"
+        subprocess.call(command, stdout=subprocess.DEVNULL, stderr=subprocess.DEVNULL, shell = True)
 
 def glitch(filename, output_name, glitch_amount = 0.7):
     """
@@ -230,6 +293,8 @@ def glitch(filename, output_name, glitch_amount = 0.7):
                        save_all=True,
                        duration=DURATION,
                        loop=LOOP)
+
+""" 4. FUNCTIONS THAT EDIT GIFS """
 
 def painting(filename, output_name, brush_width = 1, gradient = "scharr"):
     """
@@ -311,7 +376,7 @@ def gif_reduce_frames(filename, output_name, skip_frames=2):
     command = "gifsicle -U " + filename + " `seq -f \"#%g\" 0 " + str(skip_frames) + " " + num_frames + "` -o " + output_name
     os.system(command)
 
-""" UTILITY FUNCTIONS """
+""" --- UTILITY FUNCTIONS --- """
 
 def copy(filename, output_name):
     """
@@ -372,6 +437,15 @@ def callback_on_one(callback, filename, input_num, output_num, *args, **kwargs):
     output_name = get_output_filename(filename, output_num)
     callback(filename, output_name, *args, **kwargs)
 
+def acceptable_extension(filename):
+    """
+    Returns whether filename has acceptable (can be manipulated) extension/file type.
+    """
+    
+    return filename.endswith(
+        (".png", ".gif", ".qt", ".mov", ".mp4")
+    )    
+
 def callback_all(callback, input_num, output_num, *args, **kwargs):
     """
     For all files in input_path starting with input_num, executes callback on them,
@@ -380,13 +454,16 @@ def callback_all(callback, input_num, output_num, *args, **kwargs):
     Meant to be used as a wrapper around the other image manipulation functions.
     Since it lets you operate on image files in bulk.
     """
+    
     print("\n--- Callback:", callback.__name__, "| Input num:", input_num, "| Output num:", output_num, "---")
-    gather_files = [x for x in natsorted(os.listdir(input_path)) if (x.endswith(".png") or x.endswith(".gif"))]
+    
+    gather_files = [x for x in natsorted(os.listdir(input_path)) if acceptable_extension(x)]
+    
     for filename in gather_files:
         full_filename = os.path.join(input_path, filename)
         callback_on_one(callback, full_filename, input_num, output_num, *args, **kwargs)
 
-""" MAIN """
+""" --- MAIN --- """
 
 if __name__ == "__main__":
